Move to another directory; Fix imports
